+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `deploy add status` command
 - `deploy trigger` command
 
+### Changed
+
+- deprecate `deploy` command
+
 ## [v0.14.0] - 2024-07-25
 
 ### Added